adjust package name (#2699)
